requirements.txt
python-telegram-bot[persistence]
python-telegram-bot[job-queue]
gspread
oauth2client
pandas
python-dotenv